apps/data-pipeline/pyproject.toml
[project]
name = "reelix-data-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "reelix-core",
  "qdrant-client",
  "httpx>=0.27.0",
  "nltk>=3.9.0",
  "rank_bm25",
  "sentence_transformers",
  "joblib",
  "pandas",
  "sqlalchemy[postgresql]",
  "psycopg[binary]",
  "python-dotenv",
  "tqdm",
]

[tool.uv.sources]
"reelix-core" = { path = "../../packages/python", editable = true }
